requirements.txt
accelerate==1.7.0
dataclasses-json==0.6.7
datasets==4.4.2
faiss-cpu==1.11.0.post1
fastapi==0.115.9
huggingface-hub==0.36.0
hydra-core==1.3.2
ipython==8.37.0
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
python-dotenv==1.1.0
PyYAML==6.0.2
rank-bm25==0.2.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
starlette==0.45.3
tiktoken==0.9.0
tokenizers==0.22.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.67.1
transformers==4.52.4
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
flask==3.1.2
